requirements.txt
agate==1.6.1
altair==4.1.0
appdirs==1.4.4
appnope==0.1.0
asn1crypto==1.3.0
attrs==19.3.0
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
Babel==2.8.0
backcall==0.1.0
black==19.10b0
bleach==3.1.5
boto3==1.11.17
botocore==1.14.17
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
dbt==0.16.1
dbt-bigquery==0.16.1
dbt-core==0.16.1
dbt-postgres==0.16.1
dbt-redshift==0.16.1
dbt-snowflake==0.16.1
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
flatten-json==0.1.7
future==0.18.2
google-api-core==1.16.0
google-auth==1.15.0
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
great-expectations==0.10.12
hologram==0.0.6
idna==2.8
ijson==2.6.1
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
json-rpc==1.13.0
jsonschema==3.1.1
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
leather==0.3.3
Logbook==1.5.3
MarkupSafe==1.1.1
marshmallow==3.6.0
matplotlib==3.2.1
minimal-snowplow-tracker==0.0.2
mistune==0.8.4
more-itertools==8.3.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.4
oscrypto==1.2.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parsedatetime==2.5
parso==0.7.0
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.8.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.1
psycopg2-binary==2.8.5
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.9.7
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
pytest-logbook==1.2.0
python-dateutil==2.8.1
python-slugify==4.0.0
pytimeparse==1.1.8
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
regex==2020.5.14
requests==2.22.0
retrying==1.3.3
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
snowflake-connector-python==2.2.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzlocal==2.1
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
zipp==3.1.0